[feat] 리소스 소유 검증문 추가  (#397)
* test: 포트폴리오 삭제 권한 테스트 추가

* test: 포트폴리오 다수 제거 서비스의 권한 예외 테스트 추가

* test: 포트폴리오 종목 조회시 권한 예외 케이스 추가

* feat: 매입 이력 리소스 소유 권한 검증문 추가

* test: 매입 이력 추가시 권한 검증문 추가

* test: 매입이력 수정 및 삭제 서비스의 권한 검증문 추가

* test: 포트폴리오 차트 조회에 리소스 소유 검증문 추가

* feat: 권한 검증문 추가

* test: 포트폴리오 알림 상태 변경의 권한 검증 테스트 추가

* feat: 종목 지정가 단일 제거 서비스엣 권한 검증문 추가

* test: 종목 지정가 단일 제거 권한 테스트 추가

* feat: 회원 알림 서비스에 권한 검증문 추가

* test: 회원 알림의 권한 관련 테스트 추가

* feat: 관심종목 서비스에 권한 검증문 추가

* test: 관심종목 권한 테스트 추가

+++ src/main/java/codesquad/fineants/domain/holding/service/PortfolioHoldingService.java
@@ -105,9 +105,10 @@ public class PortfolioHoldingService {
 	}
 
 	@Transactional
-	public PortfolioStockDeleteResponse deletePortfolioStock(Long portfolioHoldingId) {
+	public PortfolioStockDeleteResponse deletePortfolioStock(Long portfolioHoldingId, Long memberId) {
 		log.info("포트폴리오 종목 삭제 서비스 : portfolioHoldingId={}", portfolioHoldingId);
 
+		validateHasAuthorization(List.of(portfolioHoldingId), memberId);
 		purchaseHistoryRepository.deleteAllByPortfolioHoldingIdIn(List.of(portfolioHoldingId));
 
 		int deleted = portfolioHoldingRepository.deleteAllByIdIn(List.of(portfolioHoldingId));
@@ -171,6 +172,7 @@ public class PortfolioHoldingService {
 
 	public PortfolioHoldingsResponse readPortfolioHoldings(Long portfolioId) {
 		Portfolio portfolio = findPortfolio(portfolioId);
+		validateHasAuthorization(portfolio);
 		PortfolioDetailResponse portfolioDetail = portfolioDetailFactory.createPortfolioDetailItem(portfolio);
 		List<PortfolioHoldingItem> portfolioHoldingItems = portfolioHoldingDetailFactory.createPortfolioHoldingItems(
 			portfolio);
@@ -188,6 +190,7 @@ public class PortfolioHoldingService {
 
 	public PortfolioChartResponse readPortfolioCharts(Long portfolioId, LocalDate currentLocalDate) {
 		Portfolio portfolio = findPortfolio(portfolioId);
+		validateHasAuthorization(portfolio);
 		PortfolioDetails portfolioDetails = PortfolioDetails.from(portfolio);
 		List<PortfolioPieChartItem> pieChartItems = pieChart.createBy(portfolio);
 		List<PortfolioDividendChartItem> dividendChartItems = dividendChart.createBy(portfolio, currentLocalDate);
